sidecars/learning/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
PyYAML==6.0.2
httpx==0.27.2
anthropic==0.39.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
rich==13.9.4
